pyproject.toml
[project]
name = "dbt_partitions_components"
requires-python = ">=3.10,<3.15"
version = "0.1.0"
dependencies = [
    "dagster==1.12.15",
    "dagster-dbt>=0.28.15",
    "dbt-core>=1.11.6",
    "dbt-duckdb>=1.10.1",
    "openpyxl>=3.1.0",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "dagster-webserver",
    "dagster-dg-cli",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
force-include = { "pyproject.toml" = "pyproject.toml" }

[tool.dg]
directory_type = "project"

[tool.dg.project]
root_module = "dbt_partitions_components"
registry_modules = [
    "dbt_partitions_components.components.*",
]


